+++ frontend/src/app/modules/page/admin/presentation/scheduled-registry-processes/scheduled-registry-processes.component.spec.ts
@@ -36,7 +36,7 @@ describe('ScheduledRegistryProcessesComponent', () => {
 
   it('should render data', async () => {
     await renderRegistryProcessesComponent();
-    await waitFor(() => expect(screen.getByText('1/1/22')).toBeInTheDocument());
-    await waitFor(() => expect(screen.getByText('2/1/22')).toBeInTheDocument());
+    await waitFor(() => expect(screen.getByText('1/1/00')).toBeInTheDocument());
+    await waitFor(() => expect(screen.getByText('2/1/00')).toBeInTheDocument());
   });
 });