+++ frontend/src/app/modules/shared/assembler/parts.assembler.ts
@@ -46,7 +46,8 @@ export class PartsAssembler {
       customerPartId: part.customerPartId,
       qualityType: part.qualityType || QualityType.Ok,
       productionDate: new CalendarDateModel(part.manufacturingDate),
-      children: part.childDescriptions.map(child => child.id),
+      children: part.childDescriptions.map(child => child.id) || [],
+      parents: part.parentDescriptions?.map(parent => parent.id) || [],
       shouldHighlight: part.underInvestigation || false,
       van: part.van || '--',
     };