+++ frontend/src/app/modules/shared/modules/relations/model/relations.model.ts
@@ -33,6 +33,7 @@ export interface TreeElement {
 
   state?: State;
   children?: string[];
+  parents?: string[];
 }
 
 export interface TreeStructure {
@@ -47,9 +48,15 @@ export interface TreeStructure {
 
 export interface TreeData {
   id: string;
+  mainId: string;
   treeId?: string;
   r?: number;
   defaultZoom?: number;
   openDetails?: (data: TreeStructure) => void;
-  updateChildren?: (data: TreeStructure) => void;
+  updateChildren?: (data: TreeStructure, direction: TreeDirection) => void;
+}
+
+export enum TreeDirection {
+  LEFT = 'LEFT',
+  RIGHT = 'RIGHT',
 }