+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added possibility to create an alert within my parts page, marked with WIP
 - Added additional irs request on irs-service to also include SingleLevelUsageAsBuilt param
 - Added owner column to asset table
+- Added upstream visibility for parts
 
 ### Changed
 - Updated spring-boot:core from 6.0.6 to 6.0.8 for cve-2023-20863